renamed list.tsx to be the observation panel, as thats its main output
