+++ src/main/webui/src/observations/observationPanel.tsx
